--- source/avoiding/data/init.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" ?>
-<root>
-<ret>0.336438370838</ret>
-<mtx1>[[ 626.89759615    0.          313.42588245]
- [   0.          625.28933872  223.56730617]
- [   0.            0.            1.        ]]</mtx1>
-<mtx2>[[ 624.99957837    0.          308.42990634]
- [   0.          623.60677573  216.01348773]
- [   0.            0.            1.        ]]</mtx2>
-<dist1>[[-0.42895884  0.26627144  0.0029957   0.00161685  0.        ]]</dist1>
-<dist2>[[-0.41575495  0.21015202  0.00121335  0.0015364   0.        ]]</dist2>
-<R>[[ 0.99980635 -0.01840128  0.00697517]
- [ 0.0182764   0.9996787   0.01756326]
- [-0.00729612 -0.01743238  0.99982142]]</R>
-<T>[[-33.8444782 ]
- [ -0.31494245]
- [  1.27729195]]</T>
-<E>[[  2.10464454e-02   1.27139136e+00   3.37319621e-01]
- [ -1.03011139e+00   6.13493710e-01  -3.38473437e+01]
- [  3.03673848e-01   3.38393993e+01   5.92222760e-01]]</E>
-<F>[[  2.02451698e-06   1.22613278e-04  -7.70545464e-03]
- [ -9.93106341e-05   5.92976195e-05  -2.02778722e+00]
- [  3.90850303e-02   1.98904663e+00   1.00000000e+00]]</F>
-</root>
