+++ source/implementation/avoiding/camera_testimg.py
@@ -3,7 +3,7 @@ import cv2
 import numpy as np
 
 import sys
-sys.path.append("../config")
+sys.path.append("../../config")
 import common as common_config
 
 H, W = (common_config.BINCAP_HEIGHT, common_config.BINCAP_WIDTH)
@@ -14,7 +14,7 @@ if __name__=='__main__':
     cap1 = cv2.VideoCapture(1)
     cap1.set(cv2.CAP_PROP_FRAME_WIDTH, W);
     cap1.set(cv2.CAP_PROP_FRAME_HEIGHT, H);
-    dst = np.zeros((H,3*W,3), np.uint8)
+    dst = np.zeros((H,2*W,3), np.uint8)
     i = 0
     while True:
         ret0, frame0 = cap0.read()
@@ -26,8 +26,8 @@ if __name__=='__main__':
         if cv2.waitKey(1) & 0xFF==ord('a'):
             i += 1
             print i
-            cv2.imwrite('data\\test\\0x'+str(i)+'.jpg', frame0)
-            cv2.imwrite('data\\test\\1x'+str(i)+'.jpg', frame1)
+            cv2.imwrite('data/test0/'+str(i)+'.jpg', frame0)
+            cv2.imwrite('data/test1/'+str(i)+'.jpg', frame1)
         if cv2.waitKey(1) & 0xFF==ord('q'):
             break
     cv2.destroyAllWindows()
