+++ source/implementation/avoiding/data/init.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" ?>
+<root>
+<ret>0.329533114841</ret>
+<mtx1>[[ 640.5771574     0.          309.08606981]
+ [   0.          637.99069615  228.93973146]
+ [   0.            0.            1.        ]]</mtx1>
+<mtx2>[[ 636.48380337    0.          306.30062307]
+ [   0.          634.42298038  221.55412114]
+ [   0.            0.            1.        ]]</mtx2>
+<dist1>[[-0.42706716  0.23626748  0.00343624  0.00435472  0.        ]]</dist1>
+<dist2>[[-0.42240799  0.20363647  0.00139647  0.00451319  0.        ]]</dist2>
+<R>[[ 0.99984684 -0.0173401   0.00236992]
+ [ 0.01732426  0.9998285   0.00654578]
+ [-0.00248302 -0.00650372  0.99997577]]</R>
+<T>[[-34.50442882]
+ [ -0.43632956]
+ [  2.06745719]]</T>
+<E>[[  0.03473376   2.06426485   0.44985211]
+ [ -1.98146552   0.26025716 -34.50849241]
+ [  0.16150111  34.50607719   0.22482443]]</E>
+<F>[[ -2.52388217e-07  -1.50605133e-05   1.43204738e-03]
+ [  1.44448230e-05  -1.90495841e-06   1.57118766e-01]
+ [ -3.86993379e-03  -1.55199874e-01   1.00000000e+00]]</F>
+</root>
