+++ source/implementation/avoiding/doXML.py
